# internal/server/web/openapi.yaml
...
openapi: 3.1.0
info:
  title: wudict
  version: "1"
  summary: A local dictionary server for MDX/MDD, StarDict, Slob, DSL and BGL files.
  description: |
    WuWeiDict serves your own dictionary files over HTTP, on your own machine.

    **Streaming.** `/api/dicts`, `/api/search` and `/api/rescan` answer with
    NDJSON: one JSON object per line, flushed the moment it is ready. Read them
    line by line; do not wait for the body to end. That is the whole point - the
    first dictionary's results arrive while the slowest one is still opening.
    Every stream opens with a `begin` line and closes with an `end` line.

    `/api/ingest` streams Server-Sent Events instead, because it reports
    progress rather than results.

    **Two surfaces.** Three read-only endpoints answer a browser extension
    cross-origin: `/api/dicts`, `/api/search` and `/res/`. Everything else is
    same-origin, is marked `internal` below, and may change between releases.

    **An access key, when the server is reachable.** On the default
    `127.0.0.1` bind there is nothing to authenticate and no key is asked for.
    On any other address - and Go's wildcard listen is dual-stack, so `0.0.0.0`
    includes every globally routable IPv6 address the machine holds - `AUTH`
    defaults to requiring one on every endpoint EXCEPT the three read-only ones
    above, which stay open to the extensions already granted them. Send it as
...
    and let the browser keep the cookie it sets. Without it: `401`, with
...
servers:
...
security: []
...
paths:
  /api/search:
    get:
      tags: [search]
      operationId: search
      summary: Search every dictionary, streaming
      description: |
        Queries the selected dictionaries concurrently and streams one line per
        dictionary as it finishes. Lines arrive in completion order, never in
        slot order: each `hit` carries `i`, its index in the `begin` line's
        `slots` array.

        A dictionary that fails, that cannot serve the requested mode, or that
        was declined by the memory cap reports that on its own line. None of
        those is a request failure, and the other dictionaries still answer.

        A search is cancelled after 30 seconds.
      parameters:
        - name: q
          in: query
          required: true
          description: The search text. Blank is a 400.
          schema: { type: string }
          example: flight
        - name: mode
          in: query
          description: |
            `exact` · `prefix` · `contains` (trigram substring over folded
            headwords) · `fts` (full text). `fuzzy` is accepted as a legacy
            alias and now means `prefix`.

            `contains` and `fts` need the dictionary to carry that index; ask
            `/api/dicts` for `caps` first, and expect `skipped` if you do not.

            A multi-word `fts` query is a PHRASE first. It is retried as a
            proximity and then as separate words only when the stricter reading
            returns nothing, and the hit says which reading answered (`rung`).
            Quotes and the operators `AND`, `OR`, `NOT` and `NEAR(a b, n)` are
            parsed when present, and an explicit query like
            `"faux ami" NOT linguistics` is answered exactly as written, without
            relaxation. Anything unparseable is read as plain words rather than
            rejected.
          schema:
            type: string
            enum: [exact, prefix, contains, fts, fuzzy]
            default: prefix
        - name: hl
          in: query
          description: |
            Mark the query's words inside each article body, wrapped in
            `<mark class="wu-hl">`. Read ONLY in `mode=fts`, where the reason a
            result matched is invisible otherwise; ignored in every other mode,
            where the match is the headword itself.

            Off unless you ask, so an article you fetch is the dictionary's own
            bytes by default. wudict's own UI sends `hl=1` from a remembered
            preference, which is why the default here is off and the default
            there is on. Marking is prefix-based and whole-token, matching what
            the full-text index matched, and is capped at 500 marks per article
            - past the cap the article is still returned complete, just no
            longer marked. `format=text` drops the marks with all other markup.
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          description: |
            `raw` is the dictionary's own HTML. `clean` keeps structure,
            emphasis and media and drops scripts, styles and presentation, at
            roughly half the bytes; it also rewrites root-absolute links such
            as `/res/…` into absolute URLs, so the payload works inside a page
            served from somewhere else. `text` is markup-free, roughly 0.4.

            Ask for `clean` unless you render the article with the dictionary's
            own stylesheet.
          schema:
            type: string
            enum: [raw, clean, text]
            default: raw
        - name: dict
          in: query
          description: >
            `all`, one dictionary id, or a comma-separated list of ids in the
            order you want them back. Ids come from /api/dicts. A list where no
            id resolves is a 404; unknown ids inside a resolving list are
            skipped. Naming exactly one dictionary bypasses the memory cap,
            because a search of one is a direct demand.
          schema: { type: string, default: all }
          example: oxford,webster
        - name: 'n'
          in: query
          description: Maximum results per dictionary.
          schema: { type: integer, minimum: 1, default: 20 }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/search?q=flight&mode=prefix&format=clean&n=5'

            # -N is the whole point: without it curl buffers, and the lines that
            # were sent one at a time arrive together at the end.
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SearchBegin'
                  - $ref: '#/components/schemas/SearchHit'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","i":0,"slots":[{"dict":"oxford","name":"oxford"},{"dict":"webster","name":"webster"}]}
                {"t":"hit","i":1,"dict":"webster","name":"Webster's Revised Unabridged","results":[{"Headword":"flight","Body":"<div>…</div>"}]}
                {"t":"hit","i":0,"dict":"oxford","name":"Oxford Advanced Learner's","results":[]}
                {"t":"end","i":0}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: searchPreflight
      summary: CORS preflight
      description: >
        Answers 204 with the extension grant. The extension's own requests are
        CORS-simple and never preflight; this exists so a future one does not
        meet a 405 that explains nothing.
      responses:
        '204': { description: Preflight accepted. }
  /api/dicts:
    get:
      tags: [search]
      operationId: listDicts
      summary: List every dictionary, streaming
      description: >
        `begin` carries the total immediately, from registry ids alone, so a
        client can show "0 of 105" before any dictionary is opened. Rows follow
        in completion order.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -sN 'http://127.0.0.1:6888/api/dicts'

            # just the ids, which is what /api/search wants
            curl -sN 'http://127.0.0.1:6888/api/dicts' | jq -r 'select(.t=="dict") | .dict.id'
      responses:
        '200':
          description: The NDJSON stream. One object per line.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
              example: |
                {"t":"begin","total":2}
                {"t":"dict","dict":{"id":"oxford","name":"Oxford Advanced Learner's","format":"mdx","path":"/Users/me/Dicts/oald.mdx","entries":184000,"caps":{"Exact":true,"Prefix":true,"Contains":false,"FTS":false}}}
                {"t":"end"}
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: dictsPreflight
      summary: CORS preflight
      responses:
        '204': { description: Preflight accepted. }
  /res/{dict}/{path}:
    get:
      tags: [search]
      operationId: getResource
      summary: One file from inside a dictionary
      description: |
        Images, stylesheets, scripts and audio, as the article references them.

        `path` is the whole remainder of the URL and may contain slashes, which
        an OpenAPI path parameter cannot express: send it unencoded, exactly as
        the article asks for it.

        Speex audio is transcoded to WAV and answered as `audio/wav`. A `.spx`
        that cannot be transcoded, or one supplied in a `res/` override folder,
        is answered raw as `audio/ogg`. Files from a `res/` folder are served
        uncached, so an edit takes effect on reload.
      parameters:
        - name: dict
          in: path
          required: true
          description: Dictionary id, from /api/dicts.
          schema: { type: string }
        - name: path
          in: path
          required: true
          description: The resource path inside the dictionary, subfolders included.
          schema: { type: string }
          example: audio/flight__gb.mp3
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # the path after the id is the article's own path, slashes and all
            curl -s 'http://127.0.0.1:6888/res/oxford/audio/flight__gb.mp3' -o flight.mp3
      responses:
        '200':
          description: The file, with the type the dictionary declares.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No such dictionary, or no such file inside it.
        '500':
          $ref: '#/components/responses/ServerError'
    options:
      tags: [search]
      operationId: resourcePreflight
      summary: CORS preflight
      parameters:
        - name: dict
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Preflight accepted. }
...
  /api/rescan:
    get:
      tags: [internal]
      operationId: rescan
      x-internal: true
      summary: Re-scan the dictionary folders, then list
      description: >
        Reads the dictionary folders again, warms the registry, and then answers
        exactly as /api/dicts does. Use it after adding files on disk.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # picks up files added to the dictionary folders, then lists
            curl -sN 'http://127.0.0.1:6888/api/rescan' | head -1
      responses:
        '200':
          description: The /api/dicts stream, after the rescan.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/ingest:
    get:
      tags: [internal]
      operationId: ingest
      x-internal: true
      summary: Prepare a dictionary, or add an index to it
      description: |
        Brings one dictionary to the feature state you ask for, streaming
        Server-Sent Events:

            event: progress   data: {"done":N,"total":M}
            event: done       data: {…a /api/dicts row…}
            event: error      data: {"error":"…"}

        Each switch is a state, not a verb: a parameter you leave out keeps
        whatever that feature is now, so toggling one thing cannot strip
        another. Progress is emitted at most every 200 ms.
      parameters:
        - name: dict
          in: query
          required: true
          description: Dictionary id.
          schema: { type: string }
        - name: contains
          in: query
          description: Trigram substring index over headwords. `0` or `false` turns it off.
          schema: { type: string, enum: ['0', '1', 'true', 'false'] }
        - name: fts
          in: query
          description: Full-text index over article text - the largest one.
          schema: { type: string, enum: ['0', '1', 'true', 'false'] }
        - name: media
          in: query
          description: Pack the dictionary's media into media.db.
          schema: { type: string, enum: ['0', '1', 'true', 'false'] }
        - name: full
          in: query
          description: Legacy. Any value means full text plus media.
          deprecated: true
          schema: { type: string }
        - name: level
          in: query
          description: Legacy. `headwords` turns full text off.
          deprecated: true
          schema: { type: string, enum: [headwords] }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # add the substring index to one dictionary and watch it build
            curl -sN 'http://127.0.0.1:6888/api/ingest?dict=oxford&contains=1'

            # leave a switch out and it keeps its current state; this only packs media
            curl -sN 'http://127.0.0.1:6888/api/ingest?dict=oxford&media=1'
      responses:
        '200':
          description: The event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/library:
    get:
      tags: [internal]
      operationId: listLibrary
      x-internal: true
      summary: List the prepared dictionaries on disk
      description: >
        Everything in the library folder, whether or not it is currently
        enrolled. Listing is not consent: reading this never adds anything to
        the registry.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s 'http://127.0.0.1:6888/api/library' | jq '.dir, .count'
      responses:
        '200':
          description: The library.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dir: { type: string, description: The library folder. }
                  count: { type: integer }
                  useCached:
                    type: boolean
                    description: Whether prepared dictionaries whose source file is gone are listed.
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/LibEntry' }
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [internal]
      operationId: removeLibrary
      x-internal: true
      summary: Delete one dictionary
      description: |
        Removes the prepared folder, the original files, or both. Both default
        to on, which is "remove this dictionary".

        **Irreversible.** Files are unlinked, not moved to a Trash or Recycle
        Bin, and nothing here can be undone.

        From the machine running wudict this is always allowed - it is your
        library and your disk. From a browser on another machine it follows
        `ALLOW_REMOTE_DELETE` (default off); with that off, the remote caller
        gets a 403 and `canDelete: false` in `GET /api/config`.
      parameters:
        - name: dict
          in: query
          required: true
          description: Dictionary id.
          schema: { type: string }
        - name: prepared
          in: query
          description: Delete the prepared library folder. `0` keeps it.
          schema: { type: string, default: '1' }
        - name: source
          in: query
          description: Delete the original dictionary files. `0` keeps them.
          schema: { type: string, default: '1' }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # remove one dictionary: prepared folder and original files
            curl -s -X DELETE 'http://127.0.0.1:6888/api/library?dict=oxford'

            # free the indexes only, keeping the original file
            curl -s -X DELETE 'http://127.0.0.1:6888/api/library?dict=oxford&source=0'

            # There is no dry run here - `wudict rm` is the one that previews.
      responses:
        '200':
          description: What was removed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Removal' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: >
            Deleting from another machine is off (ALLOW_REMOTE_DELETE = "0").
            Do it on the machine running wudict, or turn the setting on.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: >
            Dropping the source while keeping the prepared folder would leave a
            dictionary that exists but is listed nowhere. Enable cached
            dictionaries first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/lemmas:
    get:
      tags: [internal]
      operationId: listLemmas
      x-internal: true
      summary: List installable and installed lemma data
      description: >
        One row per language, merging the published catalogue (LEMMA_URL) with
        what is present in LEMMA_DIR, plus the state of any download running
        right now. Always answers 200: what is installed is the half of the
        answer that needs no network, and an unreachable catalogue is reported
        in `error` rather than by failing.
      parameters:
        - name: refresh
          in: query
          description: >
            Re-read the catalogue instead of using the cached copy. The
            catalogue is otherwise cached for ten minutes, so a page polling
            during a download does not ask its host once a second.
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s 'http://127.0.0.1:6888/api/lemmas' |
              jq -r '.languages[] | select(.installed) | .code'
      responses:
        '200':
          description: The catalogue merged with the folder.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dir: { type: string, description: LEMMA_DIR. }
                  url: { type: string, description: The catalogue this lists. }
                  enabled:
                    type: boolean
                    description: >
                      False when MORPH_CACHE is 0, in which case nothing
                      installed here will ever be consulted by a search.
                  cacheSize:
                    type: integer
                    description: How many language packs may be resident at once.
                  attribution:
                    type: string
                    description: ATTRIBUTION.txt published beside the catalogue (ODbL).
                  error:
                    type: string
                    description: Why the catalogue could not be read, if it could not.
                  shadowed:
                    type: array
                    description: Files the lemmatizer ignores because another supplies the same language.
                    items: { type: string }
                  languages:
                    type: array
                    items: { $ref: '#/components/schemas/LemmaLanguage' }
    post:
      tags: [internal]
      operationId: installLemmas
      x-internal: true
      summary: Install one language's lemma data
      description: |
        Starts a download and returns at once; follow it with `GET /api/lemmas`.
        The download runs in the server, not in this request, so it survives the
        page being reloaded, closed, or rotated away on Android.

        The catalogue is never taken from the caller: only the configured
        LEMMA_URL is fetched. A second call for a language already downloading
        joins the running download instead of starting another.

        On success the language is searchable immediately - no restart, because
        the lemma folder is re-indexed in place.
      parameters:
        - name: code
          in: query
          required: true
          description: Language code or English name ("pl", "polish").
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s -X POST 'http://127.0.0.1:6888/api/lemmas?code=pl'
      responses:
        '202':
          description: The download is running.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  state: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '502':
          description: The catalogue could not be read, so nothing can be installed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [internal]
      operationId: removeLemmas
      x-internal: true
      summary: Delete one language's lemma data
      description: >
        Removes every file in LEMMA_DIR supplying this language, not only the
        one in use, and re-indexes the folder. Deleting English falls back to
        the pack compiled into the binary rather than losing the language.
        Unlike dictionary removal this is offered on every platform: these are
        files wudict downloaded, not files the user brought.
      parameters:
        - name: code
          in: query
          required: true
          description: Language code or English name.
          schema: { type: string }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s -X DELETE 'http://127.0.0.1:6888/api/lemmas?code=pl'
      responses:
        '200':
          description: What was removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  builtin: { type: boolean }
                  removed:
                    type: array
                    items: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: That language is being downloaded right now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/setup:
    get:
      tags: [internal]
      operationId: setup
      x-internal: true
      summary: Validate a dictionary folder, and optionally save it
      description: >
        Drives the first-run page. Without `save` it only reports what the
        folder holds, which is what live typing needs. With `save=1` it switches
        the registry over and writes DICT_DIR. Repeat `path` for several
        folders; `found` is then the deduplicated total across all of them,
        because overlapping folders must not be counted twice.
      parameters:
        - name: path
          in: query
          description: A dictionary folder. Repeatable. `~` is expanded.
          schema: { type: string }
        - name: save
          in: query
          description: Any value persists the folders to the config file.
          schema: { type: string }
        - name: useCached
          in: query
          description: >
            `1` enrolls the previously prepared dictionaries and persists
            USE_CACHED; `0` turns that off; absent leaves it alone.
          schema: { type: string, enum: ['0', '1'] }
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # what this folder holds, without changing anything
            curl -s 'http://127.0.0.1:6888/api/setup?path=~/Dictionaries'

            # adopt two folders and write them to the config file
            curl -s 'http://127.0.0.1:6888/api/setup?path=~/Dictionaries&path=/Volumes/Ext/Dicts&save=1'
      responses:
        '200':
          description: >
            The result. A folder that does not exist, holds nothing, or cannot
            be read reports that in `error` with a 200: this answers a keystroke,
            and a half-typed path is not a failed request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  responses:
    BadRequest:
      description: A parameter is missing or unknown.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such dictionary id, or no such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: The dictionary could not be opened, or the stream could not start.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    StreamEnd:
      type: object
      description: The last line of any NDJSON stream.
      properties:
        t: { const: end }
        i:
          type: integer
          description: Always 0 here. The search stream carries the field on every line.
      required: [t]
    SearchBegin:
      type: object
      description: >
        The slot layout, sent before any dictionary is opened. Draw the empty
        result areas from it at once; each real name arrives with its hit.
      properties:
        t: { const: begin }
        i: { type: integer, description: Always 0 here. }
        slots:
          type: array
          items:
            type: object
            properties:
              dict: { type: string, description: Dictionary id. }
              name: { type: string, description: The id again; the display name arrives with the hit. }
      required: [t, slots]
...